ugc_service/app/requirements.txt
aiohttp==3.8.6
aiokafka==0.8.1
asyncpg==0.28.0
fastapi==0.103.2
gunicorn==20.1.0
orjson==3.9.1
pydantic==2.4.1
pytest==7.4.0
pytest-asyncio==0.21.1
python-dotenv==1.0.0
python-multipart==0.0.6
uvicorn==0.23.2
Werkzeug==2.3.6
pyJWT==2.8.0
loguru==0.7.2
pymongo==4.6.0
pydantic-settings==2.0.3
pydantic[email]
motor==3.3.1
